pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bird-listener"
version = "0.1.0"
description = "Real-time bird sound detection and analysis using BirdNET-Analyzer"
readme = "readme.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Andrew", email = "andrew@seatrain.net"}
]
keywords = ["birds", "audio-analysis", "birdnet", "wildlife-monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "librosa",
    "tflite-runtime",
    "asyncio",
    "uuid7",
    "pydub",
    "keras_tuner",
    "pyalsaaudio",
    "numpy<2.0",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
    "black",
    "flake8",
    "mypy",
]

[project.scripts]
bird-listener = "recordAnalyse:main"

[tool.setuptools]
py-modules = ["recordAnalyse"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
